templates: show meta of articles on the index page as well
In case readtime is configured, that is shown before the article content
on both the article and the index page.

The actual meta of an article (especially its date) is probably a much
more useful information, yet one had to navigate to the article page to
see it.

With this change, the readtime and meta is in sync, they are both shown
on the index and article pages in a consistent way.